Revamp to previews on hover + easy menu navigation
All the view inside stuff looked great, the settings. Thank you. let's change it so it says View Inside [Enter]  **[v]** (; replace the v with an ascii down arrow and the whole thing is bolded ;). to indicate that you can press down or enter. Also, when I did view inside it didn't display the instructions inside the module despite it being in the overall tree.

Please fix the bugs but also keep all fucntionality - and refactor so the overlal code is shorter without losing anything just consolidate things that are the same

Thank you. Please make these changes

[= Changes to make [ Starting now description in numbered list [

@1@= When you open settings, you don't need to open into another settings button and move them all down there as something else highlighted. You can always keep the buttons on the main heading if  we are opened into one of the sub-branches like settings.

@2@ I want to return to the main agent view when I click on Smile Composer at any time. It acts the same as an escape button, moving us out into the main view.

@3@ Remove left activating movement mode. Make left just go back out of the module or instruction. So it's like a universal back button like escape. Instead there will be a move button before the settings button after the go into button (with [Enter] etc)

@4@ When going into an instruction (pressing right) and opening the menu, make the default menu open.

@5@ This means now for every button there will be some default open. For the View Inside, this will be a text preview of every instruction or module inside the module. nothing you can navigate to, at the same size. This will appear and is not part of navigation by defalt whenever we move to an item. Once we move right the menu will switch to the settings without needing to press down. Once we are in the instruction zone the entire prompt will be visible as text instead of going deeper into the module as it is an instruction not a module. So for instructions the raw text is there. That raw text has inside of it the editable parts in different colours, red green and blue, and inside the settings those red green and blue texts correspond to the three different fields.

The default preview text on hover needs to be indented in appropriately, at the moment it is all the way left aligned.

Move doesn't appear to work anymore. After pressing right once, Move is now the next button. When I'm hovering over that button,  my up and down now needs to be in move down and be moving it up and down in the stack until i move off move mode  or press enter

The settings need to be indented 1x.